+++ Maze.Core/Objects/Point.cs
@@ -2,10 +2,10 @@
 {
     public class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(decimal x, decimal y)
         {
             this.X = x;
             this.Y = y;
